application/requirements-api.txt
fastapi~=0.110.1
uvicorn~=0.21.1
websockets==12.0
tabulate~=0.9.0
boto3~=1.33.4
psycopg2-binary==2.9.9
SQLAlchemy==1.4.52
opensearch-py==2.4.2
PyMySQL==1.1.1
python-dotenv~=1.0.0
plotly~=5.18.0
cryptography==42.0.4
langchain~=0.1.11
langchain-core~=0.1.30
sqlparse~=0.4.2
pandas==2.0.3
openpyxl
starrocks==1.0.6
clickhouse-sqlalchemy==0.2.6
sagemaker
python-jose
sqlalchemy-redshift~=0.8.14
numpy==1.26.4
pyhive==0.7.0
thrift==0.20.0
thrift-sasl==0.4.3